调整接口注释,修复两个BUG。

+++ SOUI/src/helper/SMenuEx.cpp
@@ -930,7 +930,7 @@ void SMenuEx::RunMenu(HWND hRoot)
 
         if (bMsgQuit)
         {
-            PostQuitMessage(msg.wParam);
+            PostQuitMessage((int)msg.wParam);
             break;
         }
     }
